kcm: move sec key parser to separate file so it can be shared
